start generalizing task
